Added argument validation and command line return status to generator and reflector command line tools.

+++ dotNET/Dubrovnik.Tools/Dubrovnik.Reflector/Program.cs
@@ -2,6 +2,7 @@
 using System.IO;
 using System.Linq.Expressions;
 using System.Reflection;
+using System.Runtime.Remoting.Messaging;
 using System.Text;
 using System.Xml.Linq;
 using Dubrovnik.Reflector.Command;
@@ -9,7 +10,7 @@ using Dubrovnik.Tools;
 
 namespace Dubrovnik.Reflector {
 	class Program {
-		private static void Main(string[] args)
+		private static int Main(string[] args)
 		{
 			BindingFlags bindingFlags = BindingFlags.Default;
 
@@ -30,8 +31,8 @@ namespace Dubrovnik.Reflector {
 				}
 				else
 				{
-					Console.WriteLine("Invalid binding flags.");
-					return;
+					Console.WriteLine("ERROR: Invalid binding flags.");
+					return -1;
 				}
 
 				if (options.Verbose == "true")
@@ -41,33 +42,48 @@ namespace Dubrovnik.Reflector {
 				}
 
 			} else {
-				Console.WriteLine("Command line arguments failed to parse");
-				return;
+				Console.WriteLine("Error: Command line arguments failed to parse");
+				return -1;
 			}
 
 			try {
-					
-					// load assembly
-					var inputFileName = options.InputFile;
-					Assembly assembly = Assembly.LoadFrom(inputFileName);
 
-					// parse the assembly via reflection into XML
-					AssemblyParser assemblyParser = new AssemblyParser();
-					var xml = assemblyParser.ParseAssembly(assembly, inputFileName, bindingFlags);
-
-					// output file name 
-					string xmlFileName = Path.GetFileNameWithoutExtension(inputFileName) + ".xml";
-					string outputFileName = Path.Combine(options.OutputPath, xmlFileName);
-
-					// Write xml output overwriting existing files
-					// .NET default string encoding is Unicode
-					File.WriteAllText(outputFileName, xml, Encoding.Unicode);
+				// input file must exist
+				if (!File.Exists(options.InputFile)) {
+					Console.WriteLine("ERROR: Input file does not exist.");
+					return -1;
 				}
-				catch (Exception e)
-				{
-						Console.WriteLine("An Exception Occurred : {0}", e.GetType().ToString());
+
+				// output path must exist
+				if (!Directory.Exists(options.OutputPath)) {
+
+					Console.WriteLine("ERROR: Output path does not exist.");
+					return -1;
 				}
+	
+				// load assembly
+				var inputFileName = options.InputFile;
+				Assembly assembly = Assembly.LoadFrom(inputFileName);
+
+				// parse the assembly via reflection into XML
+				AssemblyParser assemblyParser = new AssemblyParser();
+				var xml = assemblyParser.ParseAssembly(assembly, inputFileName, bindingFlags);
+
+				// output file name 
+				string xmlFileName = Path.GetFileNameWithoutExtension(inputFileName) + ".xml";
+				string outputFileName = Path.Combine(options.OutputPath, xmlFileName);
+
+				// Write xml output overwriting existing files
+				// .NET default string encoding is Unicode
+				File.WriteAllText(outputFileName, xml, Encoding.Unicode);
+			}
+			catch (Exception e)
+			{
+				Console.WriteLine("ERROR: An Exception Occurred : {0}", e.ToString());
+				return -1;
 			}
 
+			return 0;
 		}
 	}
+}
